+++ src/main/java/top/wsure/dao/ManageMapper.java
@@ -3,13 +3,13 @@ package top.wsure.dao;
 import top.wsure.entity.Manage;
 
 public interface ManageMapper {
-    int deleteByPrimaryKey(Integer manageId);
+    int deleteByPrimaryKey(Long manageId);
 
     int insert(Manage record);
 
     int insertSelective(Manage record);
 
-    Manage selectByPrimaryKey(Integer manageId);
+    Manage selectByPrimaryKey(Long manageId);
 
     int updateByPrimaryKeySelective(Manage record);
